# Conflicts:
#	src/main/java/org/team1/keyduck/common/util/ValidationErrorMessage.java

+++ src/main/java/org/team1/keyduck/auction/dto/response/AuctionReadAllResponseDto.java
@@ -16,7 +16,7 @@ public class AuctionReadAllResponseDto {
     private Long startPrice;
     private Long currentPrice;
     private Long immediatePurchasePrice;
-    private int biddingUnit;
+    private Long biddingUnit;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime auctionStartDate;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -27,7 +27,7 @@ public class AuctionReadAllResponseDto {
 
 
     private AuctionReadAllResponseDto(Long auctionId, AuctionKeyboardDto keyboard, String title,
-            Long startPrice, Long currentPrice, Long immediatePurchasePrice, int biddingUnit,
+            Long startPrice, Long currentPrice, Long immediatePurchasePrice, Long biddingUnit,
             LocalDateTime auctionStartDate, LocalDateTime auctionEndDate,
             AuctionStatus auctionStatus, Long winnerId, String winnerName) {
         this.auctionId = auctionId;
